backend/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
python-dateutil
httpx
python-dotenv
pypdf
chromadb
